🎨 Decouple markdown renderer from blog slug json
